Modul 6; opg 1

+++ 06_modul/ef-modul6/Model/Board.cs
@@ -3,11 +3,10 @@ namespace Model
 {
     public class Board
     {
-
         public Board()
         {
         }
-        public long BoardId { get; set; }
+        public int BoardId { get; set; }
         public List<Todo> Todos { get; set; }
     }
 }
